+++ packages/orq-rc/RELEASES.md
@@ -10659,3 +10659,13 @@ Based on:
 - [python v4.15.0-rc.18] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.18 - packages/orq-rc
+
+## 2026-09-03 12:41:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.19] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.19 - packages/orq-rc